Baseline Model
For now, our baseline model predicts a binary version of SCREENING_DISP_CODE using a decision tree.  After we confirm whether the target variable should be whether an arrestee is rearrested or whether an arrestee is convincted, we will update the model.  This is just to get started working with the data and modeling.

Precision is ~75% and recall is very low, ~10%.